apps/ai-worker/requirements.txt
fastapi
uvicorn
redis
sqlalchemy
psycopg2-binary
langchain-google-genai
google-generativeai
python-dotenv
twilio
